+++ app/Http/Controllers/CategoryController.php
@@ -14,6 +14,10 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        session_start();
+        if(!session()->get('name')){
+            return redirect('/');
+        }
         $categories = Category::paginate(10);
         return view('category.index',compact('categories'));
     }
@@ -25,6 +29,10 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        session_start();
+        if(!session()->get('name')){
+            return redirect('/');
+        }
         return view('category.create');
     }
 
@@ -36,6 +44,10 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        session_start();
+        if(!session()->get('name')){
+            return redirect('/');
+        }
         $request->validate([
             'name' => 'required'
         ]);
@@ -68,6 +80,10 @@ class CategoryController extends Controller
      */
     public function edit($id)
     {
+        session_start();
+        if(!session()->get('name')){
+            return redirect('/');
+        }
         $category = Category::find($id);
         return view('category.edit',compact('category'));
     }
@@ -81,6 +97,10 @@ class CategoryController extends Controller
      */
     public function update(Request $request, $id)
     {
+        session_start();
+        if(!session()->get('name')){
+            return redirect('/');
+        }
         $request->validate([
             'name' => 'required'
         ]);
@@ -102,6 +122,10 @@ class CategoryController extends Controller
      */
     public function destroy(Request $request)
     {
+        session_start();
+        if(!session()->get('name')){
+            return redirect('/');
+        }
         //dd($request);
         $category = Category::find($request->id);
         $category->movies()->detach();
